Start implementing splint and RATS cleanups.
* src/manage.h (append_to_credentials_username): Change third arg to
gsize.
(current_server_task): Annotate as null.

* src/manage.c: Define g_free with annotation for splint.  Include
string and strings.  Define PREFIX.  Declare functions and vars static.
Move all local vars to beginnings of blocks.  Annotate global vars as null.
(append_to_credentials_username, append_to_credentials_password): Append
with append_string.
(report_path_task_name): Change type of path2_length to size_t.
(print_tasks): Use standard ternary format.
(grow_tasks): Add "RATS: ignore" to realloc.
(grow_description): Add "RATS: ignore" to realloc.  Clear new memory.

* src/file.c: Declare fchdir for splint.  Move all local variable
declarations to beginnings of blocks.
(rmdir_recursively): Cast variables to quiet splint.

* src/string.h (append_string, free_string_var): Drop arg names.
(strip_space): Annotate return as shared.


git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-manager@2938 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/manage.h
@@ -47,10 +47,10 @@ void
 free_credentials (credentials_t*);
 
 void
-append_to_credentials_username (credentials_t*, const char*, int);
+append_to_credentials_username (credentials_t*, const char*, gsize);
 
 void
-append_to_credentials_password (credentials_t*, const char*, int);
+append_to_credentials_password (credentials_t*, const char*, gsize);
 
 int
 authenticate (credentials_t);
@@ -144,7 +144,7 @@ extern unsigned int tasks_size;
 /**
  * @brief The task currently running on the server.
  */
-extern task_t* current_server_task;
+extern /*@null@*/ task_t* current_server_task;
 
 int
 task_id_string (task_t*, const char **);
